% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit_imp}
\alias{fit_imp}
\title{Models fitted using the imputed data}
\format{
A tibble with 11 rows and 2 variables:
\describe{
  \item{ind_pulse}{Indicator for missing pulse}
  \item{ind_rr}{Indicator for missing blood pressure}
  \item{ind_spo2}{Indicator for missing SpO2}
  \item{ind_gcs}{Indicator for missing GCS}
  \item{ind_rhythm}{Indicator for missing rhythm}
  \item{set}{Name of the set}
  \item{mod}{Fitted model used for getting model fits}
}
}
\usage{
fit_imp
}
\description{
Contains models for each of the missing variable combinations
}
\keyword{datasets}
