% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_wrangle_variable.R
\name{wrangle_variable}
\alias{wrangle_variable}
\title{wrangle_variable:  Wrangle a single input variable for the champ calculator}
\usage{
wrangle_variable(var_name, df_in, df_definitions)
}
\arguments{
\item{var_name}{Name of the variable to be wrangled}

\item{df_in}{Input data frame}

\item{df_definitions}{Variable definitions gotten from the 
variable_definitions.xlsx}
}
\value{
Vector containing the wrangled variable to be used in the calculate_champ-function
}
\description{
A utils function
}
\examples{

# files in the root directory
if (FALSE) {
df_defintions <- readxl::read_excel("variable_definitions_example.xlsx", sheet = 1)
df_in <- readxl::read_excel("example_data.xlsx")

# wrangle data
c_age <- wrangle_variable("Age (years)", df_in, df_defintions)
c_pulse <- wrangle_variable("Heart rate (bpm)", df_in, df_defintions)
c_bp <- wrangle_variable("Systolic blood pressure (mmHg)", df_in, df_defintions)
c_spo2 <- wrangle_variable("Oxygen saturation (\%)", df_in, df_defintions)
c_time <- wrangle_variable("Time from alarm to HEMS arrival (minutes)", df_in, df_defintions)
c_gcs <- wrangle_variable("Glasgow Coma Scale", df_in, df_defintions)
c_sex <- wrangle_variable("Patient sex", df_in, df_defintions)
c_crhythm <- wrangle_variable("Cardiac rhythm", df_in, df_defintions)
c_med_facility <- wrangle_variable("Medical facility or nursing home", df_in, df_defintions)
c_vehicle <- wrangle_variable("HEMS vehicle", df_in, df_defintions)
c_code <- wrangle_variable("Patient group", df_in, df_defintions)

# calculate risk
calculate_champ(rr = c_bp,
               pulse = c_pulse,
               spo2 = c_spo2,
               gcs = c_gcs,
               time_from_alarm = c_time,
               cardiac_rhythm = c_crhythm,
               age = c_age,
               medical_facility = c_med_facility,
               vehicle_ground_unit = c_vehicle,
               sex_man = c_sex,
               code = c_code)

}
}
